refactor: remove picocolors from shared
