+++ packages/compat/plugin-swc/package.json
@@ -31,6 +31,7 @@
     "@swc/helpers": "0.5.11",
     "core-js": "~3.37.1",
     "lodash": "^4.17.21",
+    "picocolors": "^1.0.1",
     "semver": "^7.6.2"
   },
   "devDependencies": {
