+++ packages/compat/webpack/src/progress/helpers/index.ts
@@ -1,4 +1,4 @@
 export * from './bus';
 export * from './bar';
-export * from './type';
+export * from './types';
 export * from './percentage';
