+++ packages/compat/webpack/src/progress/helpers/nonTty.ts
@@ -1,5 +1,5 @@
 import { logger } from '@rsbuild/core';
-import { color } from '@rsbuild/shared';
+import color from 'picocolors';
 
 export function createNonTTYLogger() {
   let prevPercentage = 0;
