+++ packages/core/src/cli/commands.ts
@@ -1,6 +1,7 @@
 import { existsSync } from 'node:fs';
-import { type RsbuildMode, color } from '@rsbuild/shared';
+import type { RsbuildMode } from '@rsbuild/shared';
 import { type Command, program } from 'commander';
+import color from 'picocolors';
 import { isEmptyDir } from '../helpers';
 import { logger } from '../logger';
 import { init } from './init';
