+++ packages/core/src/plugins/fileSize.ts
@@ -4,8 +4,8 @@
  */
 import fs from 'node:fs';
 import path from 'node:path';
-import { color } from '@rsbuild/shared';
 import type { PrintFileSizeOptions, Stats, StatsAsset } from '@rsbuild/shared';
+import color from 'picocolors';
 import { CSS_REGEX, HTML_REGEX, JS_REGEX } from '../constants';
 import { logger } from '../logger';
 import type { RsbuildPlugin } from '../types';
