+++ packages/core/src/plugins/html.ts
@@ -1,11 +1,6 @@
 import fs from 'node:fs';
 import path, { isAbsolute } from 'node:path';
-import {
-  type MinifyJSOptions,
-  castArray,
-  color,
-  deepmerge,
-} from '@rsbuild/shared';
+import { type MinifyJSOptions, castArray, deepmerge } from '@rsbuild/shared';
 import type {
   HTMLPluginOptions,
   HtmlConfig,
@@ -13,6 +8,7 @@ import type {
   NormalizedEnvironmentConfig,
 } from '@rsbuild/shared';
 import type { EntryDescription } from '@rspack/core';
+import color from 'picocolors';
 import {
   reduceConfigsMergeContext,
   reduceConfigsWithContext,
