+++ packages/core/src/server/helper.ts
@@ -3,7 +3,6 @@ import net from 'node:net';
 import type { Socket } from 'node:net';
 import os from 'node:os';
 import { join, relative } from 'node:path';
-import { color } from '@rsbuild/shared';
 import type {
   NormalizedConfig,
   OutputStructure,
@@ -11,6 +10,7 @@ import type {
   Routes,
   RsbuildEntry,
 } from '@rsbuild/shared';
+import color from 'picocolors';
 import { DEFAULT_DEV_HOST, DEFAULT_PORT } from '../constants';
 import { isFunction } from '../helpers';
 import { logger } from '../logger';
