+++ packages/plugin-check-syntax/package.json
@@ -31,6 +31,7 @@
     "acorn": "^8.12.0",
     "browserslist-to-es-version": "^1.0.0",
     "htmlparser2": "9.1.0",
+    "picocolors": "^1.0.1",
     "source-map": "^0.7.4"
   },
   "devDependencies": {
